[Cache] moved classes
